Fixed JSX to work via tailwind config and changed file locations around...
