+++ src/App.js
@@ -1,43 +1,30 @@
-import "./App.css";
 // importing components from react-router-dom package
 import {
   BrowserRouter as Router,
-  Switch,
   Route,
-  Redirect,
+  Routes,
 } from "react-router-dom";
-  
-import Home from "./pages/home";
-import About from "./pages/about";
-import Checkout from "./pages/checkout";
-import Basicpackage from "./pages/basicpackage";
-import Header from "./pages/header";
-import Footer from "./pages/footer";
-  
+
+import Home from "./Routes/Home";
+import About from "./Routes/About";
+import Checkout from "./Routes/Checkout";
+import BasicPackage from "./Routes/BasicPackage";
+import Navbar from "./Components/Navbar";
+import Footer from "./Components/Footer";
+
 function App() {
   return (
     <>
-      {/* This is the alias of BrowserRouter i.e. Router */}
-      <Router>
-      <div>
-        <Header />
-         <hr />
-            <Switch>
-
-              <Route exact from="/" to="/home" component={Home} />
-                
-              <Route path="/about" component={About} />
-                
-              <Route path="/checkout" component={Checkout} />
-                
-              <Route path="/basicpackage" component={Basicpackage} />
-              <Redirect from ="/" to="/home" />
-            </Switch>
-          <Footer />
-        </div>
-      </Router>
+      <Navbar />
+      <Routes>
+        <Route path="/" to="/Home" element={<Home />} />
+        <Route path="/About" element={<About />} />
+        <Route path="/Checkout" element={<Checkout />} />
+        <Route path="/BasicPackage" element={<BasicPackage />} />
+      </Routes>
+      <Footer />
     </>
   );
 }
-  
+
 export default App;
