+++ src/Routes/About.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const About = () => {
+  return (
+    <div>About</div>
+  )
+}
+
+export default About
